+++ Cargo.toml
@@ -14,5 +14,5 @@ simplelog = "0.12"
 sha2 = "0.10"
 native-dialog = "0.7.0"
 serde = { version = "1.0.216", features = ["derive"] }
-serde_yaml = "0.9.34+deprecated"
+serde_yaml = "0.9.34"
 reqwest = { version = "0.11", features = ["json"] }
