+++ src/tabs/about_tab.rs
@@ -1,7 +1,7 @@
 use eframe::egui;
 
+// 窗口形式显示关于内容
 pub fn show_about_window(ctx: &egui::Context, open: &mut bool) {
-    let version = env!("CARGO_PKG_VERSION"); // 从环境变量读取版本号
     egui::Window::new("关于此软件")
         .open(open)
         .resizable(false)
@@ -11,7 +11,7 @@ pub fn show_about_window(ctx: &egui::Context, open: &mut bool) {
         });
 }
 
-// 新增：直接在UI面板中显示关于内容
+// 面板形式显示关于内容
 pub fn show_about_content(ui: &mut egui::Ui) {
     let version = env!("CARGO_PKG_VERSION");
     
@@ -47,4 +47,9 @@ pub fn show_about_content(ui: &mut egui::Ui) {
     ui.heading("鸣谢:");
     ui.label("egui - 一个简单、快速、高度可移植的即时模式 GUI 库");
     ui.hyperlink_to("egui 官方网站", "https://github.com/emilk/egui");
+}
+
+// 处理关于标签页面
+pub fn handle_about_tab(ui: &mut egui::Ui) {
+    show_about_content(ui);
 }
