+++ src/tabs/mod.rs
@@ -1,0 +1,3 @@
+pub mod about_tab;
+pub mod clear_tab;
+pub mod ai_ui_tab;
