+++ src/utils.rs
@@ -27,6 +27,7 @@ use std::fs;
 use std::path::Path;
 use sha2::{Digest, Sha256};
 
+#[allow(dead_code)]
 pub fn hash_file(path: &Path) -> Result<String, std::io::Error> {
     let mut file = fs::File::open(path)?;
     let mut hasher = Sha256::new();
@@ -34,6 +35,7 @@ pub fn hash_file(path: &Path) -> Result<String, std::io::Error> {
     Ok(format!("{:x}", hasher.finalize()))
 }
 
+#[allow(dead_code)]
 pub fn compare_dirs_hash(source: &Path, target: &Path) -> Result<bool, std::io::Error> {
     let source_hashes: Vec<_> = fs::read_dir(source)?
         .map(|entry| hash_file(&entry?.path()))
